regression/requirements.txt
mlflow==2.18.0
catboost==1.2.7
cloudpickle==3.1.0
ipython==8.28.0
numpy==1.26.4
pandas==2.2.3
psutil==6.1.0
scikit-learn==1.5.2
scipy==1.14.1
torch==2.5.1
tqdm==4.67.1
xgboost==2.1.3
fastapi==0.115.0
uvicorn==0.15.0
python-dotenv==1.0.0
boto3==1.21.0
pydantic==1.10.7